squash commits
This is the 1st commit message:

Create POC of working with audio

Create 3 POCs:
- One POC of playing the sound as it is recorded
- One POC of recording audio to file (using pickle)
- One POC of playing audio from file (using pickle)

Each of the POCs work, but none of them work in the same way we want the
final components to work. I do not know how robust each of these POC is,
and if they will work in any computer without altering them (for example
if the default device is not the one you want, it will not work).

I still don't know how to specify in the recording only POC that I want
the data to go into a numpy array.

We still need to try a POC of passing the audio over TCP (and decide how
to pass it - maybe just struct.pack?)

This is the commit message #2:

 Start working on Net Player

Found out the 16KHz recording and playing works

Started work (very early stages) on the Net Player file, which will
contain a POC for playing a recording in real time from a different
process than the one that recorded it.

This is the commit message #3:

Add input visualizer

Added a visualizer to the network player.
currently only visualizing directly from microphone works.
none of the other modes work: microphone to net, net to speakers, net to visualizer.
Also, currently the only arguments supported are the modes, and everything else is hard coded.

This is the commit message #4:

Add non blocking option to visualizer

In order to support the visualize from net mode, I needed to add a non blocking mode to the visualizer.
In non blocking mode, you can decide how long you want to let the visualizer block.

This is needed, as the network receiver needs run time in the main thread in order to receive data,
unlike the recorder receiver that uses another thread in order to record in the background.

This is the commit message #5:

Make visualize from net and record to net modes work

Make the microphone to network and network to visualizer modes work.
The microphone to network does not yet manage to close when the socket closes.

Still need to make the network to speaker mode work.

This is the commit message #6:

Make recorder mode close when the socket is closed

This is the commit message #7:

Make everything close nicely when socket sender runs out of time

Change both the socket sender and socket receiver to close without
exceptions when the socket sender runs out of time and closes its
socket.

We probably won't use this feature so much, but it's good practice
to not throw an exception when a socket closes and instead close
everything nicely and return gracefully.

This is the commit message #8:

Add initial class diagram

This is the commit message #9:

Add TODO list to readme file

This is the commit message #10:

Fix TODO list to use actual checkbox

This is the commit message #11:

Move Writer and Reader classes and implementations to src folder

Rename the Listener and Receiver to Writer and Reader respectively.
Create writer and reader folders, and create a file for each abstract
class and for each implementation.

This is the commit message #12:

Add option to run visualizer and microphone as classes of the library

Added a temporary run option to the library, to allow the option to
run two simple clasees to check that after moving prom POC to real
classes everything still works.
Currently only the microphone reader and audio visualizer work in
this mode.

This is the commit message #13:

Add speaker player writer

Add a writer that plays sound received from the reader to the speakers.

The speaker player is still not documented, and blocks forever when
wait() is called.

This is the commit message #14:

Change speaker player stream from with to the more general __enter__

To support the option to make the stream non blocking, change the
way to open the stream from a "with" statement to an "__enter__" call.

This is the commit message #15:

Make speaker player non-blocking

This is the commit message #16:

Create a processor writer

This is the commit message #17:

Create a splitter processor

Create a processor that send the data to multiple writers.

This is the commit message #18:

Test splitter processor with multiple writers

test the splitter (in run.py) with multiple visualize writers and multiple speaker players.

This is the commit message #19:

Add pipeline processor

This is the commit message #20:

add multiplier processor

Add a processor that can increase or decrease the volume of the input data

This is the commit message #21:

Add dict parser to run script

This is the commit message #22:

Make run script load the classes dict from a yaml file

This is the commit message #23:

Add example of socket to main yaml run

This is the commit message #24:

Add option to make the wait function of SocketWriter not block

Change the timeout argument of SocketWriter, so that when you put a
value that is not None in it, the wait funciton will only block for this
long and then will return to the reader.

Added a complex example that uses sockets to the yamls for the run script,
and updated the readme with instuctions about how to run it.

This is the commit message #25:

Create file reader and writer

Create a file writer and a file reader.
Add yaml files to test the recording to files and playing from files.

This is the commit message #26:

Add and update plantUML documentation

This is the commit message #27:

Add sample sound files

move yaml files from src to test
add sample sound files and yaml config to play them
add support for sample size of 2 in AudioVisualizer

This is the commit message #28:

Add finalize function to all writers and processors

This is the commit message #29:

Fix bug of speaker writer terminating before getting any data

This is the commit message #30:

Add requirements.txt

This is the commit message #31:

Add conda environment yaml file

This is the commit message #32:

Update requirements.txt and install notes

This is the commit message #33:

Change requirements.txt to only include direct dependencies NOT TESTED

This is the commit message #34:

Add dccrn processor

Add a processor working with an implementatoin of dccrn to reduce noise.
Add a test config that includes this processor.

This is the commit message #35:

Update requirements.txt with requirements for DCCRN

This is the commit message #36:

Remove POC folder and POC files

This is the commit message #37:

Add support for samplesize of 2 in dccrn_processor

This is the commit message #38:

Minor improvements to readme and error output

This is the commit message #39:

Add option to read and write wav files in addition to raw files

This is the commit message #40:

Implement overlaping windows in dccrn_processor to reduce small window size artifacts

This is the commit message #41:

Improve dccrn overlapping window to have more fine control over the overlapping function

This is the commit message #42:

Add utils module to convert samples from raw data to sample array and back

This is the commit message #43:

Make run script only import objects it should use according to the config file parameter

This is the commit message #44:

Make the run script easier to run

This is the commit message #45:

Make the run script easier to use from importing code

Make the run script easier to use when importing it, to make it possible
to expand the library and implement readers, writers, and processors outside
of the library and still use the run script.

This is the commit message #46:

rename real_time_noise_processing to real_time_audio_processing

This is the commit message #47:

Add documentation to all the objects and functions

This is the commit message #48:

Fix indentation error

This is the commit message #49:

Change abstract classes to be ABC abstract classes

This is the commit message #50:

remove test wav files